Fix two spec failures regarding Document#[]. Fix the same bug in Element#at.
Document#[] is meant to return either an element with a given id,
the first element with a given css path or the first element with a
given xpath path whichever comes first.

The previous implementation assumed that Element#css and Element#xpath
return nil if a path is invalid and that was an incorrect assumption:
they may throw a JS exception.

The new implementation tries to catch such an exception so that the
method works as advertised.

Same with Element#at except that this one doesn't have a spec.