Make Document#ready? return true when either "complete" or "interactive"
As per [1], there's another state that wasn't taken into account and that
is "interactive" which is between a DOM load and a full load. We expect
Document#ready? (DOM load) to always be true when Document#ready calls back.
This was a corner case when it was false.

This one caused a delayed test failure.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Document/readyState