Fix another minor error or two on a test page.
This one was caused, again, by a too weak exception catch.

A JS::Error is not a subclass of a StandardError (just Exception).
Expressions in the form of "x rescue nil" are equivalent of
"begin x rescue; nil end" which obviously only catch StandardErrors.

This commit makes a test page free from exceptions and green like a
cuke.