Implement Native::Wrapper::Singleton and migrate the library to use it.
This may not be an obvious improvement and it certainly looks destructive.
I checked the whole library and tried to test it where possible. It seems
like it works well and doesn't break any assumptions when applied carefully.

When implementing Iframe#content_window I had a realization that every next
call to this function created a new Ruby object. I wondered if that's the
best idea. I think it's also how jQuery encapsulates things. But we are not
jQuery.

I noticed that DOM::Node includes a #== override. That's nice, but Window
is not a subclass of DOM::Node. There should have been some better way
than adding a #== method in Native::Wrapper.

I got inspired by the way we do .new dispatching in DOM::Node. Why not return
there an existing object if it was already created. This way our Ruby DOM
should kind of reflect the native DOM better.

I thought about storing a map of references in some global variable, but
I realized it would cause issues with a GC. Storing a reference to a Ruby
object inside a JS object seems like a good compromise.

The name Singleton usually means a class with a single instance. That may
not be the best name here, but I couldn't find a better one.

This change should offer some memory improvements under certain circumstances
and under certain it won't.