Refactor DOM::Element::Scroll. Remove Window::Scroll.
This commit basically refactors the DOM::Element::Scroll module.
This module was broken before, because there were multiple conflicting
definitions of a #to function. I retained the classic one adding
some syntactic sugar like $window.scroll.to :bottom. The other one
got renamed to #into_view.

My refactor integrated this code with Window::Scroll and removed
this module afterwards.

Inside this commit there are also minor caching changes for Window.

After having implemented all that I noticed, that actually there are
two DOM interfaces for Element.scroll* and Window.scroll* with a
differing browser support. The thing we do is extracting a root
Element from a Window so we don't need to use the Window interface
at all (and it should be supported well cross-browser - and it's
actually what we did partially with the Window::Scroll implementation
before this commit).