Rename Native::Wrapper::Singleton to Browser::NativeCachedWrapper.
As per @elia's suggestion, because, in his own words:

"So, the thing is, this looks like something that should end up in
opal's stdlib. If by any chance the final version has a different
api that opal-browser's version it will give us some trouble in
reconciling the two, that's why I'd like to keep it in a namespace"

This commit also isolates ClassMethods to be located a level higher
as Browser::NativeCachedWrapperClassMethods so as not to pollute the
namespace of the including modules.

This commit also renames the JS property name to "$$opal_native_cached"
as per Opal current guidelines.