Implement Browser.restricted?. Handle Restricted objects.
Now, this commit may be quite controversial and if some of you find
it too controversial, I would gladly hear from you.

I worked on Iframe#content_window, which sometimes returns a weird
object. It's a DOM node, but it's "Restricted", you can't edit its
properties, but it has a few features that you can use, like a
postMessage method.

None of the Native::Wrapper or Browser::NativeCachedWrapper could
work with such a node, because they checked a DOM node with a
Kernel#native? method - which checks for `#@native.$$class` - this
triggers a SecurityError. Therefore if we could catch such an
exception... maybe we could treat this kind of object in a special
way?

Below is a sample code which can create such an object:

```
z = DOM { iframe sandbox: "" }
z.append_to($document.body)
z.content_window
```

This workaround uses Browser::NativeCachedWrapper. If it detects,
that a native object that is about to be wrapped is restricted, it
skips all other checks and creates a limited object. It works,
because we migrated all DOM to use NativeCachedWrapper. Of course,
such objects aren't cached anymore (well, it's not anymore - they
never were, because they never worked), but it may be possible to
cache them with a different approach, maybe further justifying the
location of this workaround...